Implement reload when product is added to store.
Been trying this for 3-4 hours now. Finally freaking did it, here's how:
Add product/package appears as bottomsheet of scaffold. Using await, the program waits for the bottom sheet to be closed and then reloads the state if the result returned is when the context is popped is "true". If not, it doesn't reload anything.

(#9)