Add animations to Message class.
This was an attempt to prevent the infinite Message calls, but I also added animations.
I struggled to add a check to see if the latestTime that a Message was added was newer than the last and therefore cancel the clear() method call accordingly. This way, it ignores the duration of the latest snackbar message and removes it according to the timeout of an older one.
Probably should use a list to accomplish that, but this seems fine for now.

Did some minor refactoring, too.
Additionally, added color for when a store item is selected.